refactor(engine): reuse code to determine taken flows
Introduces a new analyzer (classes that can analyze, but not change
state), to analyze sequence flows. Specifically to reuse the code to
determine which incoming sequence flows have been taken.
