+++ engine/src/main/java/io/zeebe/engine/processing/bpmn/behavior/BpmnBehaviorsImpl.java
@@ -70,7 +70,6 @@ public final class BpmnBehaviorsImpl implements BpmnBehaviors {
             streamWriter,
             zeebeState.getKeyGenerator(),
             stateBehavior,
-            deferredRecordsBehavior,
             new ProcessEngineMetrics(zeebeState.getPartitionId()),
             stateTransitionGuard,
             processorLookup,
